+++ lambdas/shared/src/common/models/errors.py
@@ -12,9 +12,10 @@ class Severity(str, Enum):
 class Code(str, Enum):
     forbidden = "forbidden"
     not_found = "not-found"
-    invalid = "invalid"
-    server_error = "exception"
+    invalid = "invalid or missing access token"
+    server_error = "internal server error"
     invariant = "invariant"
+    incomplete = "parameter-incomplete"
     not_supported = "not-supported"
     duplicate = "duplicate"
     # Added an unauthorized code its used when returning a response for an unauthorized vaccine type search.
@@ -23,6 +24,12 @@ class Code(str, Enum):
 
 @dataclass
 class UnauthorizedError(RuntimeError):
+    response: dict | str
+    message: str
+
+    def __str__(self):
+        return f"{self.message}\n{self.response}"
+
     @staticmethod
     def to_operation_outcome() -> dict:
         msg = "Unauthorized request"
@@ -36,6 +43,12 @@ class UnauthorizedError(RuntimeError):
 
 @dataclass
 class UnauthorizedVaxError(RuntimeError):
+    response: dict | str
+    message: str
+
+    def __str__(self):
+        return f"{self.message}\n{self.response}"
+
     @staticmethod
     def to_operation_outcome() -> dict:
         msg = "Unauthorized request for vaccine type"
@@ -49,6 +62,12 @@ class UnauthorizedVaxError(RuntimeError):
 
 @dataclass
 class UnauthorizedVaxOnRecordError(RuntimeError):
+    response: dict | str
+    message: str
+
+    def __str__(self):
+        return f"{self.message}\n{self.response}"
+
     @staticmethod
     def to_operation_outcome() -> dict:
         msg = "Unauthorized request for vaccine type present in the stored immunization resource"
@@ -61,10 +80,48 @@ class UnauthorizedVaxOnRecordError(RuntimeError):
 
 
 @dataclass
+class TokenValidationError(RuntimeError):
+    response: dict | str
+    message: str
+
+    def __str__(self):
+        return f"{self.message}\n{self.response}"
+
+    @staticmethod
+    def to_operation_outcome() -> dict:
+        msg = "Missing/Invalid Token"
+        return create_operation_outcome(
+            resource_id=str(uuid.uuid4()),
+            severity=Severity.error,
+            code=Code.invalid,
+            diagnostics=msg,
+        )
+
+
+@dataclass
+class ConflictError(RuntimeError):
+    response: dict | str
+    message: str
+
+    def __str__(self):
+        return f"{self.message}\n{self.response}"
+
+    @staticmethod
+    def to_operation_outcome() -> dict:
+        msg = "Conflict"
+        return create_operation_outcome(
+            resource_id=str(uuid.uuid4()),
+            severity=Severity.error,
+            code=Code.duplicate,
+            diagnostics=msg,
+        )
+
+
+@dataclass
 class ResourceNotFoundError(RuntimeError):
     """Return this error when the requested FHIR resource does not exist"""
 
-    resource_type: str
+    resource_type: str | None
     resource_id: str
 
     def __str__(self):
@@ -113,6 +170,25 @@ class UnhandledResponseError(RuntimeError):
             resource_id=str(uuid.uuid4()),
             severity=Severity.error,
             code=Code.server_error,
+            diagnostics=self.__str__(),
+        )
+
+
+@dataclass
+class BadRequestError(RuntimeError):
+    """Use when payload is missing required parameters"""
+
+    response: dict | str
+    message: str
+
+    def __str__(self):
+        return f"{self.message}\n{self.response}"
+
+    def to_operation_outcome(self) -> dict:
+        return create_operation_outcome(
+            resource_id=str(uuid.uuid4()),
+            severity=Severity.error,
+            code=Code.incomplete,
             diagnostics=self.__str__(),
         )
 
@@ -201,28 +277,23 @@ class IdentifierDuplicationError(RuntimeError):
         )
 
 
-def create_operation_outcome(resource_id: str, severity: Severity, code: Code, diagnostics: str) -> dict:
-    """Create an OperationOutcome object. Do not use `fhir.resource` library since it adds unnecessary validations"""
-    return {
-        "resourceType": "OperationOutcome",
-        "id": resource_id,
-        "meta": {"profile": ["https://simplifier.net/guide/UKCoreDevelopment2/ProfileUKCore-OperationOutcome"]},
-        "issue": [
-            {
-                "severity": severity,
-                "code": code,
-                "details": {
-                    "coding": [
-                        {
-                            "system": "https://fhir.nhs.uk/Codesystem/http-error-codes",
-                            "code": code.upper(),
-                        }
-                    ]
-                },
-                "diagnostics": diagnostics,
-            }
-        ],
-    }
+@dataclass
+class ServerError(RuntimeError):
+    """Use when there is a server error"""
+
+    response: dict | str
+    message: str
+
+    def __str__(self):
+        return f"{self.message}\n{self.response}"
+
+    def to_operation_outcome(self) -> dict:
+        return create_operation_outcome(
+            resource_id=str(uuid.uuid4()),
+            severity=Severity.error,
+            code=Code.server_error,
+            diagnostics=self.__str__(),
+        )
 
 
 @dataclass
@@ -263,6 +334,31 @@ class RecordProcessorError(Exception):
     The diagnostics dictionary received from the Record Processor is passed to the exception as an argument
     and is stored as an attribute.
     """
+
+
+def create_operation_outcome(resource_id: str, severity: Severity, code: Code, diagnostics: str) -> dict:
+    """Create an OperationOutcome object. Do not use `fhir.resource` library since it adds unnecessary validations"""
+    return {
+        "resourceType": "OperationOutcome",
+        "id": resource_id,
+        "meta": {"profile": ["https://simplifier.net/guide/UKCoreDevelopment2/ProfileUKCore-OperationOutcome"]},
+        "issue": [
+            {
+                "severity": severity,
+                "code": code,
+                "details": {
+                    "coding": [
+                        {
+                            "system": "https://fhir.nhs.uk/Codesystem/http-error-codes",
+                            "code": code.upper(),
+                        }
+                    ]
+                },
+                "diagnostics": diagnostics,
+            }
+        ],
+    }
+
 
     def __init__(self, diagnostics_dictionary: dict):
         self.diagnostics_dictionary = diagnostics_dictionary
